requirements.txt
groq==0.4.2
python-dotenv==1.0.0
requests==2.31.0
pandas==2.1.4
matplotlib==3.7.2
seaborn==0.12.2
numpy==1.24.3 